requirements.txt
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.34.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msrest==0.7.1
numpy==2.3.0
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
